fix all warnings

+++ include/latte/blob.h
@@ -26,7 +26,7 @@ class Blob : public Noncopyable {
 
   inline string shape_string() const {
     ostringstream stream;
-    for (int i = 0; i < shape_.size(); ++i) {
+    for (size_t i = 0; i < shape_.size(); ++i) {
       stream << shape_[i] << " ";
     }
     stream << "(" << count_ << ")";
@@ -73,11 +73,11 @@ class Blob : public Noncopyable {
   }
 
   inline int offset(const vector<int> &indices) const {
-    CHECK_LE(indices.size(), num_axes());
+    CHECK_LE(static_cast<int>(indices.size()), num_axes());
     int offset = 0;
     for (int i = 0; i < num_axes(); ++i) {
       offset *= shape(i);
-      if (indices.size() > i) {
+      if (static_cast<int>(indices.size()) > i) {
         CHECK_GE(indices[i], 0);
         CHECK_LT(indices[i], shape(i));
         offset += indices[i];
